refactor(helper): extract POST serialization from HttpConnection into HttpConnectionPost
